Change coordinate stuff, rework das

+++ src/inputstate.rs
@@ -1,6 +1,6 @@
 pub struct InputState {
-    pub das_right: f64,
-    pub das_left: f64,
+    pub das_right: u64,
+    pub das_left: u64,
     pub down_frames: u64,
     pub left: bool,
     pub right: bool,
@@ -10,8 +10,8 @@ pub struct InputState {
 impl InputState {
     pub fn new() -> InputState {
         InputState {
-            das_right: 0.0,
-            das_left: 0.0,
+            das_right: 0,
+            das_left: 0,
             down_frames: 0,
             left: false,
             right: false,
